Fix keyboard shortcuts
