+++ src/components/ResultFooter.js
@@ -1,10 +1,22 @@
 import { Button } from '../../UI/Button';
 import resultStyles from '../../UI/Result/result.module.css';
 import { Draw, RoundLost, RoundWon } from '../../UI/Result';
-import React from 'react';
+import React, {useEffect} from 'react';
 import { DRAW_COMPUTER, DRAW_PLAYER, PLAYER_COMPUTER, PLAYER_USER } from '../pages/game';
 
 export function ResultFooter({roundWinner, nextRound}) {
+
+   useEffect(() => {
+      document.addEventListener('keydown', handleEnter, false);
+      return () => document.removeEventListener('keydown', handleEnter)
+   }, []);
+
+   const handleEnter = (e) => {
+      if (e.key === 'Enter') {
+         nextRound()
+      }
+   }
+
    return (
       <div className={resultStyles.resultFooter}>
          {(roundWinner === DRAW_PLAYER || roundWinner === DRAW_COMPUTER) && (
